Remove constant pool (#23)
Add a new Delta that removes the constant pool by inlining the constants to where they are referenced. From a programmer's perspective this is nicer since references to constant pool indices are hard to work with, and the advantage of using a constant pool, namely that constants can be re-used, is hardly ever used.